requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
httpx>=0.28.0
pycryptodome>=3.20.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
